[Compresion] Minor refactoring
